tumor_detection/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-tumor"
version = "0.1.0"
description = "Brain tumor segmentation and auxiliary MRI classification project."
requires-python = ">=3.10"
dependencies = [
  "kagglehub",
  "matplotlib",
  "monai",
  "nibabel",
  "numpy",
  "pandas",
  "pillow",
  "pyyaml",
  "scikit-learn",
  "tensorboard",
  "torch",
  "torchvision",
  "tqdm",
]

[project.optional-dependencies]
nnunet = ["nnunetv2"]

[tool.setuptools.packages.find]
where = ["src"]
